# perf(sdk): replace Box<dyn Client> with ClientWrapper enum