--- a/core/integration/src/quic_client.rs
+++ b/core/integration/src/quic_client.rs
@@ -18,7 +18,7 @@
 
 use crate::test_server::{ClientFactory, Transport};
 use async_trait::async_trait;
-use iggy::prelude::{Client, QuicClientConfig};
+use iggy::prelude::{Client, ClientWrapper, QuicClientConfig};
 use iggy::quic::quick_client::QuicClient;
 use std::sync::Arc;
 
@@ -29,14 +29,14 @@ pub struct QuicClientFactory {
 
 #[async_trait]
 impl ClientFactory for QuicClientFactory {
-    async fn create_client(&self) -> Box<dyn Client> {
+    async fn create_client(&self) -> ClientWrapper {
         let config = QuicClientConfig {
             server_address: self.server_addr.clone(),
             ..QuicClientConfig::default()
         };
         let client = QuicClient::create(Arc::new(config)).unwrap();
         Client::connect(&client).await.unwrap();
-        Box::new(client)
+        ClientWrapper::Quic(client)
     }
 
     fn transport(&self) -> Transport {