--- a/core/integration/src/test_server.rs
+++ b/core/integration/src/test_server.rs
@@ -53,7 +53,7 @@ pub enum IpAddrKind {
 
 #[async_trait]
 pub trait ClientFactory: Sync + Send {
-    async fn create_client(&self) -> Box<dyn Client>;
+    async fn create_client(&self) -> ClientWrapper;
     fn transport(&self) -> Transport;
     fn server_addr(&self) -> String;
 }