--- a/core/sdk/src/client_provider.rs
+++ b/core/sdk/src/client_provider.rs
@@ -16,6 +16,7 @@
  * under the License.
  */
 
+use crate::client_wrappers::client_wrapper::ClientWrapper;
 #[allow(deprecated)]
 use crate::clients::client::IggyClient;
 use crate::http::http_client::HttpClient;
@@ -174,7 +175,7 @@ pub async fn get_client(config: Arc<ClientProviderConfig>) -> Result<IggyClient,
 /// Create a `Client` for the specific transport based on the provided configuration.
 pub async fn get_raw_connected_client(
     config: Arc<ClientProviderConfig>,
-) -> Result<Box<dyn Client>, ClientError> {
+) -> Result<ClientWrapper, ClientError> {
     get_raw_client(config, true).await
 }
 
@@ -182,7 +183,7 @@ pub async fn get_raw_connected_client(
 pub async fn get_raw_client(
     config: Arc<ClientProviderConfig>,
     establish_connection: bool,
-) -> Result<Box<dyn Client>, ClientError> {
+) -> Result<ClientWrapper, ClientError> {
     let transport = config.transport.clone();
     match transport.as_str() {
         QUIC_TRANSPORT => {
@@ -191,12 +192,12 @@ pub async fn get_raw_client(
             if establish_connection {
                 Client::connect(&client).await?
             };
-            Ok(Box::new(client))
+            Ok(ClientWrapper::Quic(client))
         }
         HTTP_TRANSPORT => {
             let http_config = config.http.as_ref().unwrap();
             let client = HttpClient::create(http_config.clone())?;
-            Ok(Box::new(client))
+            Ok(ClientWrapper::Http(client))
         }
         TCP_TRANSPORT => {
             let tcp_config = config.tcp.as_ref().unwrap();
@@ -204,7 +205,7 @@ pub async fn get_raw_client(
             if establish_connection {
                 Client::connect(&client).await?
             };
-            Ok(Box::new(client))
+            Ok(ClientWrapper::Tcp(client))
         }
         _ => Err(ClientError::InvalidTransport(transport)),
     }