… core/integration/tests/cli/common/mod.rs
… core/integration/tests/sdk/producer/background.rs
… core/integration/tests/server/specific.rs
… core/sdk/src/client_wrappers/binary_personal_access_token_client.rs
… core/sdk/src/client_wrappers/binary_segment_client.rs
… core/sdk/src/client_wrappers/binary_stream_client.rs
… core/sdk/src/client_wrappers/binary_system_client.rs
… core/sdk/src/client_wrappers/binary_topic_client.rs
… core/sdk/src/client_wrappers/binary_user_client.rs
… core/sdk/src/client_wrappers/client_wrapper.rs
… core/sdk/src/client_wrappers/mod.rs
… core/sdk/src/clients/client.rs
… core/sdk/src/clients/client_builder.rs
… core/sdk/src/clients/consumer.rs
… core/sdk/src/clients/consumer_builder.rs
… core/sdk/src/clients/producer.rs
… core/sdk/src/clients/producer_builder.rs
… core/sdk/src/prelude.rs