## Summary
Replace `Box<dyn Client>` trait objects with a concrete `ClientWrapper` enum to eliminate dynamic dispatch overhead and reduce heap allocations.

By leveraging match-based delegation in `ClientWrapper`, we can eliminate dynamic dispatch and `Box` allocations for client instances. The enum is stack-allocated, compared to heap-allocated Box, saving heap memory.

The `ClientWrapper` enum currently contains Http, Tcp, Quic, and Iggy variants. Future protocol implementations should be added as new variants.

This change serves as a V2 implementation and alternative to the approach in #1912. As discussed, using pure static dispatch would alter too many use case behaviors and require users to have deeper API knowledge. With the enum wrapper, no API changes are needed, benefiting `IggyClient, IggyConsumer, IggyProducer`, and other components without changing how they are used.

Performance improvements:
- `IggyClient::from_connection_string()` execution time reduced by 3.45%
- Average byte allocation reduced from 18,626 to 18,618 bytes
- Allocation speed improved by 3.53%

## Performance test
Simple performance test against `IggyClient::from_connection_string()`. Measured its execution time and heap allocation with the following 2 scripts.
Test platform is Darwin Kernel Version 22.6.0 on arm64 architecture.

Execution time:
```
use criterion::{black_box, criterion_group, criterion_main, Criterion};
use iggy::prelude::{IggyClient};

fn bench_iggy_client_new(c: &mut Criterion) {
    c.bench_function("iggy_client_new", |b| {
        b.iter(|| {
            black_box(IggyClient::from_connection_string("iggy://iggy:iggy@localhost:8090"))
        })
    });
}

criterion_group!(benches, bench_iggy_client_new);
criterion_main!(benches);
```

Heap allocation:
```
use criterion::{black_box, criterion_group, criterion_main, Criterion};
use std::alloc::{GlobalAlloc, Layout, System};
use std::sync::atomic::{AtomicUsize, Ordering};

// Custom allocator that tracks allocations
struct TrackingAllocator;

static ALLOCATED_BYTES: AtomicUsize = AtomicUsize::new(0);
static ALLOCATION_COUNT: AtomicUsize = AtomicUsize::new(0);
static DEALLOCATED_BYTES: AtomicUsize = AtomicUsize::new(0);
static DEALLOCATION_COUNT: AtomicUsize = AtomicUsize::new(0);

unsafe impl GlobalAlloc for TrackingAllocator {
    unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
        let ptr = unsafe { System.alloc(layout) };
        if !ptr.is_null() {
            ALLOCATED_BYTES.fetch_add(layout.size(), Ordering::Relaxed);
            ALLOCATION_COUNT.fetch_add(1, Ordering::Relaxed);
        }
        ptr
    }

    unsafe fn dealloc(&self, ptr: *mut u8, layout: Layout) {
        unsafe {System.dealloc(ptr, layout) };
        DEALLOCATED_BYTES.fetch_add(layout.size(), Ordering::Relaxed);
        DEALLOCATION_COUNT.fetch_add(1, Ordering::Relaxed);
    }
}

#[global_allocator]
static GLOBAL: TrackingAllocator = TrackingAllocator;

fn reset_allocation_counters() {
    ALLOCATED_BYTES.store(0, Ordering::Relaxed);
    ALLOCATION_COUNT.store(0, Ordering::Relaxed);
    DEALLOCATED_BYTES.store(0, Ordering::Relaxed);
    DEALLOCATION_COUNT.store(0, Ordering::Relaxed);
}

fn get_allocation_stats() -> (usize, usize, usize, usize) {
    (
        ALLOCATED_BYTES.load(Ordering::Relaxed),
        ALLOCATION_COUNT.load(Ordering::Relaxed),
        DEALLOCATED_BYTES.load(Ordering::Relaxed),
        DEALLOCATION_COUNT.load(Ordering::Relaxed),
    )
}

use iggy::prelude::{IggyClient};

fn bench_iggy_client_allocation(c: &mut Criterion) {
    c.bench_function("iggy_client_allocation", |b| {
        b.iter_custom(|iters| {
            reset_allocation_counters();
            
            let start = std::time::Instant::now();
            
            for _ in 0..iters {
                let _ = black_box(IggyClient::from_connection_string("iggy://iggy:iggy@localhost:8090"));
            }
            
            let elapsed = start.elapsed();

            let (allocated, alloc_count, deallocated, dealloc_count) = get_allocation_stats();
            eprintln!("Iterations: {}", iters);
            eprintln!("Allocations per iteration: {:.2}", alloc_count as f64 / iters as f64);
            eprintln!("Bytes per iteration: {:.2}", allocated as f64 / iters as f64);
            eprintln!("---");
            
            elapsed
        })
    });
}

criterion_group!(benches, bench_iggy_client_allocation);
criterion_main!(benches);
```

Result are showned in the following, comparing to older implementation

Execution time:
```
$ cargo bench --bench iggy_client_test
   Compiling iggy v0.7.0 (/Users/vax-r/iggy/core/sdk)
    Finished `bench` profile [optimized] target(s) in 1m 02s
     Running benches/iggy_client_test.rs (/Users/vax-r/iggy/target/release/deps/iggy_client_test-22c1dc27e916a5b5)
iggy_client_new         time:   [1.1030 µs 1.1045 µs 1.1059 µs]
                        change: [-3.6387% -3.4529% -3.2761%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  11 (11.00%) low mild
  1 (1.00%) high severe
```

Heap allocation:
* Old:
```
Iterations: 86300
Allocations per iteration: 24.00
Bytes per iteration: 18626.00
```
* New:
```
Iterations: 86300
Allocations per iteration: 23.00
Bytes per iteration: 18618.00
```

## Related Issue
#1873