@marvin-hansen 2025-07-04
I think generics doesn't make much sense any longer.  You would need generics over the protocl type P and the client type T, so sometihing like Client<T<Protocol<P>> but because Rust doesn't have higher kinder types, you would need to fall back on Generic Associated Types (GATs) with trait bounds over the protocol.

However, I had this situation just a few days ago in my deep_causality project and the resulting code complexity really needs to be justified. As for the IggyClient, I believe the ClientWrapper Enum is the most sensible path forward.