@vax-r 2025-07-04
> Looks solid, appreciated all the benchmarks too! Do you think it will be still worth trying out the generic approach at some point in the future?

For `IggyClient`'s case, I don't think using generic type would be sane for now, I might change my opinion when I'm more familiar with Rust.
It would be more suitable to use generic type in those situation that we certainly know what type we're going to dealt with before runtime and we can treat them all in the same manner.

> I think generics doesn't make much sense any longer. You would need generics over the protocl type P and the client type T, so sometihing like `Client<T<Protocol<P>> ` but because Rust doesn't have higher kinded types, you would need to fall back on Generic Associated Types (GATs) with trait bounds over the protocol.
> 
> However, I had this situation just a few days ago in my deep_causality project and the resulting code complexity really needs to be justified. As for the IggyClient, I believe the ClientWrapper Enum is the most sensible path forward.

Yes I agree with @marvin-hansen . Over the past 2 weeks I've been trying to port everything onto `IggyClient<T>` , `IggyProducer<T>` and so on.

The main issue is lack of flexibility, almost zero tbh. Taking `client_provider::get_raw_client()` as example, returning `IggyClient<T>` will need different protocols to be dealt in the same way, but as you can see `HTTP_TRANSPORT` is dealt slightly different than `QUIC_TRANSPORT, TCP_TRANSPORT`. As the result, we can't just treat every `T` ( protocol type, client type, whatever ) in the same way.

Splitting things in their own specific `impl` block is a workaround, but you can see in #1912 , it will require the caller to be split in the same manner, we can't have a function that adapts through different protocols, everything needs to be protocol-sensitive and dedicated to certain one.