@marvin-hansen 2025-07-04
Thank you for doing all the important digging on the details. It really seems that generic client doesn't add anything of value. 

I want to point out though, with the current state of generics in Rust, there is no way to know upfront for sure so you really have to try to know. I had those experiments going either way before. Sometimes it worked out, sometimes it was a dead end. However, i tend to see those as "stepping stones" as Ken Stanley put it in "the myth of the objective" because regardless of the outcome you learned something valuable that informs your next step. 

I supported the exploration of a generic client a while ago because, on paper, it was adding up. Now you have proven, well, down the details it's not worth the effort. And that's okay. 

The only thing I suggest is to document the decision in the client wrapper by stating something like generic client was explored, problem X,Y, Z has been encountered, and that means A, B, C therefore it was decided to use an Enum for dispatch. 

It helps any future contributors to understand why the client is the way it is.