@spetz 2025-07-04
@marvin-hansen need to look into Coveralls report, as recently we moved it from the previous repo, but apparently, it didn't start working out of the box.